requirements.txt
flask
flask-cors
flask-limiter
python-dotenv
ultralytics
easyocr
mysql-connector-python
opencv-python-headless
numpy
